Rename and tweak
- rename VisionAttachments
- tweak how system works
